cli: move validation
